Fix following rebase conflicts
